Move Any
* `Any` doesn't deserve such a prominent location. Tuck it away.
* Enable most warnings in the Cabal file and squash them. Unused
  imports are still an issue.
